ukify: Add sign command for UKIs
We want to sign UKIs that are already in a UKI format, so we can do
offline-signing, e.g. on another computer, while keeping the rest of the
UKI, including the stub, intact.
Additionally, the cmdline can be modified, to allow inserting a
roothash, which is also only known after signing.

# Conflicts:
#	src/ukify/ukify.py